#### Get Current Location

+++ app/src/main/java/com/dicoding/storyapp/api/ApiService.kt
@@ -50,7 +50,7 @@ interface ApiService {
         @Header("Authorization") token: String,
         @Part file: MultipartBody.Part,
         @Part("description") description: RequestBody,
-        @Part("lat") lat: Float,
-        @Part("lon") lon: Float,
+        @Part("lat") lat: Double,
+        @Part("lon") lon: Double,
     ): Call<CreateStoryResponse>
 }
